requirements.txt
kafka-python
statsmodels
numpy